pyproject.toml
[project]
name = "daily-wellness-backend"
version = "0.1.0"
description = "A FastAPI backend for tracking daily wellness"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.2",
    "google-genai>=1.38.0",
    "python-dotenv>=1.1.1",
    "ruff>=0.13.0",
    "sqlalchemy>=2.0.43",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.25.0",
    "pytest-mock>=3.12.0",
    "pytest-cov>=4.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow tests",
]
